fixedfibo

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,14 +9,14 @@
 
 int main(void)
 {
-	long long int i, t1 = 1, t2 = 2, nextTerm;
+	long int i, t1 = 1, t2 = 2, nextTerm;
 
 	for (i = 1; i <= 50; ++i)
 	{
 		if (i != 50)
-			printf("%lld, ", t1);
+			printf("%ld, ", t1);
 		else
-			printf("%lld", t1);
+			printf("%ld", t1);
 		nextTerm = t1 + t2;
 		t1 = t2;
 		t2 = nextTerm;
